break the memory platform into several files
